requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.13.2
requests
azure-mgmt-apimanagement==4.0.1
azure-identity==1.16.0